pyproject.toml
[build-system]
requires = ["setuptools", "setuptools-scm"]
build-backend = "setuptools.build_meta"

[project]
name = "trans_voice"
version = "0.1"
authors = [
    {name = "LiYiMing"},
]
requires-python = ">=3.8"
license = {file = "LICENSE"}

dependencies = [
    "openai",
    "python-dotenv",
]
